Move AiVad to anomalib.video
