+++ src/anomalib/models/image/__init__.py
@@ -1,0 +1,4 @@
+"""Anomalib Image Models."""
+
+# Copyright (C) 2022 Intel Corporation
+# SPDX-License-Identifier: Apache-2.0
