+++ src/anomalib/models/video/__init__.py
@@ -1,0 +1,8 @@
+"""Anomalib Video Models."""
+
+# Copyright (C) 2022 Intel Corporation
+# SPDX-License-Identifier: Apache-2.0
+
+from .ai_vad import AiVad
+
+__all__ = ["AiVad"]
