+++ src/anomalib/models/video/ai_vad/clip/__init__.py
