+++ src/anomalib/models/video/ai_vad/clip/clip.py
