+++ src/anomalib/models/video/ai_vad/clip/model.py
