+++ src/anomalib/models/video/ai_vad/density.py
@@ -12,7 +12,8 @@ from sklearn.mixture import GaussianMixture
 from torch import nn
 
 from anomalib.metrics.min_max import MinMax
-from anomalib.models.ai_vad.features import FeatureType
+
+from .features import FeatureType
 
 
 class BaseDensityEstimator(nn.Module, ABC):
