+++ src/anomalib/models/video/ai_vad/regions.py
