+++ src/anomalib/models/video/ai_vad/torch_model.py
@@ -1,4 +1,4 @@
-"""Attribute-based Representations for Accurate and Interpretable Video Anomaly Detection.
+"""PyTorch model for AI-VAD model implementation.
 
 Paper https://arxiv.org/pdf/2212.00789.pdf
 """
@@ -7,25 +7,20 @@ Paper https://arxiv.org/pdf/2212.00789.pdf
 # SPDX-License-Identifier: Apache-2.0
 
 
-import logging
-from typing import Any
-
 import torch
-from lightning.pytorch.utilities.types import STEP_OUTPUT
+from torch import nn
 
-from anomalib.models.ai_vad.torch_model import AiVadModel
-from anomalib.models.components import AnomalyModule, MemoryBankMixin
-
-logger = logging.getLogger(__name__)
-
-__all__ = ["AiVad"]
+from .density import CombinedDensityEstimator
+from .features import FeatureExtractor
+from .flow import FlowExtractor
+from .regions import RegionExtractor
 
 
-class AiVad(MemoryBankMixin, AnomalyModule):
-    """AI-VAD: Attribute-based Representations for Accurate and Interpretable Video Anomaly Detection.
+class AiVadModel(nn.Module):
+    """AI-VAD model.
 
     Args:
-        box_score_thresh (float): Confidence threshold for bounding box predictions.
+        box_score_thresh (float): Confidence threshold for region extraction stage.
         persons_only (bool): When enabled, only regions labeled as person are included.
         min_bbox_area (int): Minimum bounding box area. Regions with a surface area lower than this value are excluded.
         max_bbox_overlap (float): Maximum allowed overlap between bounding boxes.
@@ -45,24 +40,33 @@ class AiVad(MemoryBankMixin, AnomalyModule):
 
     def __init__(
         self,
-        box_score_thresh: float = 0.7,
+        # region-extraction params
+        box_score_thresh: float = 0.8,
         persons_only: bool = False,
         min_bbox_area: int = 100,
         max_bbox_overlap: float = 0.65,
         enable_foreground_detections: bool = True,
         foreground_kernel_size: int = 3,
         foreground_binary_threshold: int = 18,
-        n_velocity_bins: int = 1,
+        # feature-extraction params
+        n_velocity_bins: int = 8,
         use_velocity_features: bool = True,
         use_pose_features: bool = True,
         use_deep_features: bool = True,
-        n_components_velocity: int = 2,
+        # density-estimation params
+        n_components_velocity: int = 5,
         n_neighbors_pose: int = 1,
         n_neighbors_deep: int = 1,
     ) -> None:
         super().__init__()
+        if not any((use_velocity_features, use_pose_features, use_deep_features)):
+            msg = "Select at least one feature type."
+            raise ValueError(msg)
 
-        self.model = AiVadModel(
+        # initialize flow extractor
+        self.flow_extractor = FlowExtractor()
+        # initialize region extractor
+        self.region_extractor = RegionExtractor(
             box_score_thresh=box_score_thresh,
             persons_only=persons_only,
             min_bbox_area=min_bbox_area,
@@ -70,7 +74,16 @@ class AiVad(MemoryBankMixin, AnomalyModule):
             enable_foreground_detections=enable_foreground_detections,
             foreground_kernel_size=foreground_kernel_size,
             foreground_binary_threshold=foreground_binary_threshold,
+        )
+        # initialize feature extractor
+        self.feature_extractor = FeatureExtractor(
             n_velocity_bins=n_velocity_bins,
+            use_velocity_features=use_velocity_features,
+            use_pose_features=use_pose_features,
+            use_deep_features=use_deep_features,
+        )
+        # initialize density estimator
+        self.density_estimator = CombinedDensityEstimator(
             use_velocity_features=use_velocity_features,
             use_pose_features=use_pose_features,
             use_deep_features=use_deep_features,
@@ -79,51 +92,43 @@ class AiVad(MemoryBankMixin, AnomalyModule):
             n_neighbors_deep=n_neighbors_deep,
         )
 
-    @staticmethod
-    def configure_optimizers() -> None:
-        """AI-VAD training does not involve fine-tuning of NN weights, no optimizers needed."""
-        return
-
-    def training_step(self, batch: dict[str, str | torch.Tensor]) -> None:
-        """Training Step of AI-VAD.
-
-        Extract features from the batch of clips and update the density estimators.
+    def forward(self, batch: torch.Tensor) -> tuple[list[torch.Tensor], list[torch.Tensor], list[torch.Tensor]]:
+        """Forward pass through AI-VAD model.
 
         Args:
-            batch (dict[str, str | torch.Tensor]): Batch containing image filename, image, label and mask
-        """
-        features_per_batch = self.model(batch["image"])
-
-        for features, video_path in zip(features_per_batch, batch["video_path"], strict=True):
-            self.model.density_estimator.update(features, video_path)
-
-    def fit(self) -> None:
-        """Fit the density estimators to the extracted features from the training set."""
-        self.model.density_estimator.fit()
-
-    def validation_step(self, batch: dict[str, str | torch.Tensor], *args, **kwargs) -> STEP_OUTPUT:
-        """Perform the validation step of AI-VAD.
-
-        Extract boxes and box scores..
-
-        Args:
-            batch (dict[str, str | torch.Tensor]): Input batch
-            *args: Arguments.
-            **kwargs: Keyword arguments.
+            batch (torch.Tensor): Input image of shape (N, L, C, H, W)
 
         Returns:
-            Batch dictionary with added boxes and box scores.
+            list[torch.Tensor]: List of bbox locations for each image.
+            list[torch.Tensor]: List of per-bbox anomaly scores for each image.
+            list[torch.Tensor]: List of per-image anomaly scores.
         """
-        del args, kwargs  # Unused arguments.
+        self.flow_extractor.eval()
+        self.region_extractor.eval()
+        self.feature_extractor.eval()
 
-        boxes, anomaly_scores, image_scores = self.model(batch["image"])
-        batch["pred_boxes"] = [box.int() for box in boxes]
-        batch["box_scores"] = [score.to(self.device) for score in anomaly_scores]
-        batch["pred_scores"] = torch.Tensor(image_scores).to(self.device)
+        # 1. get first and last frame from clip
+        first_frame = batch[:, 0, ...]
+        last_frame = batch[:, -1, ...]
 
-        return batch
+        # 2. extract flows and regions
+        with torch.no_grad():
+            flows = self.flow_extractor(first_frame, last_frame)
+            regions = self.region_extractor(first_frame, last_frame)
 
-    @property
-    def trainer_arguments(self) -> dict[str, Any]:
-        """AI-VAD specific trainer arguments."""
-        return {"gradient_clip_val": 0, "max_epochs": 1, "num_sanity_val_steps": 0}
+        # 3. extract pose, appearance and velocity features
+        features_per_batch = self.feature_extractor(first_frame, flows, regions)
+
+        if self.training:
+            return features_per_batch
+
+        # 4. estimate density
+        box_scores = []
+        image_scores = []
+        for features in features_per_batch:
+            box, image = self.density_estimator(features)
+            box_scores.append(box)
+            image_scores.append(image)
+
+        box_locations = [batch_item["boxes"] for batch_item in regions]
+        return box_locations, box_scores, image_scores
